SensorTest/trafego_rodoviario/requirements.txt
Django==4.2
psycopg2-binary==2.9.3
djangorestframework==3.14.0
django-filter==2.4.0
drf-yasg==1.21.10
black==25.1.0
django-extensions==4.1.0